Move dto interfaces to dedicated files.
