+++ src/utils/interfaces/dto/IAnswerForCreation.ts
@@ -1,0 +1,5 @@
+export interface IAnswerForCreation {
+    questionId: string;
+    value: string;
+    filePath?: string | undefined;
+}
