+++ src/utils/interfaces/dto/IAnswerForPut.ts
@@ -1,0 +1,4 @@
+export interface IAnswerForPut {
+    value: string;
+    filePath?: string | undefined;
+}
