+++ src/utils/interfaces/dto/IComment.ts
@@ -1,0 +1,5 @@
+export interface IComment {
+    id: string;
+    userName: string;
+    value: string;
+}
