+++ src/utils/interfaces/dto/ICommentForCreation.ts
@@ -1,0 +1,4 @@
+export interface ICommentForCreation {
+    answerId: string;
+    value: string;
+}
