+++ src/utils/interfaces/dto/IDetailedQuestion.ts
@@ -1,0 +1,7 @@
+import { IAnswer } from "./IAnswer";
+import { IQuestion } from "./IQuestion";
+
+export interface IDetailedQuestion extends IQuestion {
+    filePath: string;
+    answers: IAnswer[];
+}
