+++ src/utils/interfaces/dto/IPaginationMeta.ts
@@ -1,0 +1,6 @@
+export interface IPaginationMeta {
+    PageNr: number;
+    Limit: number;
+    TotalItemCount: number;
+    TotalPageCount: number;
+}
