+++ src/utils/interfaces/dto/IProblemDetail.ts
@@ -1,0 +1,5 @@
+export interface IProblemDetail {
+    status: number;
+    title: string;
+    detail: string | undefined;
+}
