+++ src/utils/interfaces/dto/IQuestionForEdit.ts
@@ -1,0 +1,8 @@
+export interface IQuestionForEdit {
+    topicId: string;
+    subjectId: string;
+    tags: string[];
+    description?: string;
+    title: string;
+    isProtected: boolean;
+}
