+++ src/utils/interfaces/dto/ISubject.ts
@@ -1,0 +1,10 @@
+import { ITeacher } from "./ITeacher";
+import { ITopic } from "./ITopic";
+
+export interface ISubject {
+    id: string;
+    name: string;
+    subjectCode?: string;
+    teachers?: ITeacher[];
+    topics: ITopic[];
+}
