+++ src/utils/interfaces/dto/ISubjectForPut.ts
@@ -1,0 +1,6 @@
+export interface ISubjectForPut {
+    id: string;
+    name: string;
+    subjectCode?: string;
+    teachers: string[];
+}
