+++ src/utils/interfaces/dto/ITag.ts
@@ -1,0 +1,4 @@
+export interface ITag {
+    id: string;
+    value: string;
+}
