+++ src/utils/interfaces/dto/ITeacher.ts
@@ -1,0 +1,5 @@
+export interface ITeacher {
+    id: string;
+    email: string;
+    userName: string;
+}
