+++ src/utils/interfaces/dto/ITopic.ts
@@ -1,0 +1,6 @@
+export interface ITopic {
+    id: string;
+    name: string;
+    isActive: boolean;
+    subjectId: string;
+}
