+++ src/utils/interfaces/dto/ITopicForCreation.ts
@@ -1,0 +1,4 @@
+export interface ITopicForCreation {
+    name: string;
+    subjectId: string;
+}
