+++ src/utils/interfaces/dto/IUser.ts
@@ -1,0 +1,6 @@
+export interface IUser {
+    id: string;
+    userName: string;
+    email: string;
+    isBlocked: boolean;
+}
